requirements.txt
yt-dlp~=2026.3.17
python-dotenv~=1.2.2
aiogram~=3.26.0